+++ datastructure/src/jvmMain/kotlin/com/sunnychung/lib/multiplatform/bigtext/core/layout/CharMeasurer.kt
@@ -1,5 +1,7 @@
 package com.sunnychung.lib.multiplatform.bigtext.core.layout
 
+import com.sunnychung.lib.multiplatform.bigtext.annotation.TemporaryApi
+
 interface CharMeasurer<S> {
 
     @Deprecated("Not maintained")
@@ -8,4 +10,7 @@ interface CharMeasurer<S> {
     fun findCharWidth(char: CharSequence, style: S? = null): Float
 
     fun findCharYOffset(char: CharSequence, style: S? = null): Float
+
+    @TemporaryApi
+    fun getRowHeight(): Float
 }