+++ CHANGELOG.md
@@ -10,6 +10,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 UI
 - New `isCacheTextLayoutResult` parameter (default to `true`) to `CoreBigTextField` to cache `androidx.compose.ui.text.TextLayoutResult` in order to optimize text rendering performance
 
+DS
+- Interface `CharMeasurer` now requires an implementation of `getRowHeight()`. This API is marked as `@TemporaryApi` and may be changed when multiple row heights would be supported.
+
+Common
+- Annotation class `@TemporaryApi` is moved from the UI module to the data structure module 
+
 
 ## UI [2.1.0] - 2025-03-09